# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TodoAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/task/list": {
            "post": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get list of tasks based on filters",
                "description": "Returns list of tasks",
                "operationId": "getTasks",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "status": {
                                        "type": "string",
                                        "example": "in:todo,done|nullable"
                                    },
                                    "priority_min": {
                                        "type": "integer",
                                        "example": "integer|min:1|max:5|nullable"
                                    },
                                    "priority_max": {
                                        "type": "integer",
                                        "example": "integer|min:1|max:5|nullable"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "string|nullable"
                                    },
                                    "order_by": {
                                        "type": "string",
                                        "example": "in:created_at,completedAt,priority|nullable"
                                    },
                                    "order_direction": {
                                        "type": "string",
                                        "example": "in:asc,desc|nullable"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 94
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 12
                                            },
                                            "parent_id": {
                                                "type": "integer",
                                                "example": null,
                                                "nullable": true
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Task 0 for UserWest Nataliaville"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Description for task 0 of UserEast Francescaton"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "todo"
                                            },
                                            "priority": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "completedAt": {
                                                "type": "string",
                                                "example": "null"
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2023-09-10T16:20:31.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2023-09-10T16:20:31.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/api/task/{task_id}": {
            "get": {
                "tags": [
                    "Tasks"
                ],
                "summary": "Get task by task_id",
                "description": "Returns task",
                "operationId": "getTaskId",
                "parameters": [
                    {
                        "name": "task_id",
                        "in": "path",
                        "description": "The ID of the task",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 94
                                            },
                                            "user_id": {
                                                "type": "integer",
                                                "example": 12
                                            },
                                            "parent_id": {
                                                "type": "integer",
                                                "example": 92,
                                                "nullable": true
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Task 0 for UserWest Nataliaville"
                                            },
                                            "description": {
                                                "type": "string",
                                                "example": "Description for task 0 of UserEast Francescaton"
                                            },
                                            "status": {
                                                "type": "string",
                                                "example": "todo"
                                            },
                                            "priority": {
                                                "type": "string",
                                                "example": "1"
                                            },
                                            "completedAt": {
                                                "type": "string",
                                                "example": "null",
                                                "nullable": true
                                            },
                                            "created_at": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2023-09-10T16:20:31.000000Z"
                                            },
                                            "updated_at": {
                                                "type": "string",
                                                "format": "datetime",
                                                "example": "2023-09-10T16:20:31.000000Z"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Bearer Token Authentication",
                "name": "Bearer",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}